+++ R/from_datetime_offset.r
@@ -7,25 +7,29 @@
 #'
 #' @param x A [datetime_offset()] object
 #' @param from A [datetime_offset()] object
+#' @param year If missing what year to assume
 #' @param month If missing what month to assume
 #' @param day   If missing what day to assume
 #' @param hour   If missing what hour to assume
 #' @param minute   If missing what minute to assume
 #' @param second   If missing what second to assume
+#' @param nanosecond   If missing what nanosecond to assume
 #' @param tz   If missing and hour offset also missing what time zone to assume
 #' @param ... Ignored
 #' @name from_datetime_offset
 #' @examples
 #'   as.Date(as_datetime_offset("2020-03-05"))
 #'   as.Date(as_datetime_offset("2020"))
-#'   as.Date(as_datetime_offset("2020"), 6, 15)
+#'   as.Date(as_datetime_offset("2020"), month = 6, day = 15)
 #'
-#'   as.nanotime
+#'   as.nanotime(as_datetime_offset(Sys.time()))
+#'   as.POSIXct(as_datetime_offset(Sys.time()), tz = "GMT")
 NULL
 
 #' @rdname from_datetime_offset
 #' @export
-as.Date.datetime_offset <- function(x, month = 1, day = 1, ...) {
+as.Date.datetime_offset <- function(x, year = 1970L, month = 1L, day = 1L, ...) {
+    year(x) <- update_missing(year(x), year)
     month(x) <- update_missing(month(x), month)
     day(x) <- update_missing(day(x), day)
 
@@ -41,28 +45,47 @@ setOldClass("datetime_offset")
 #' @rdname from_datetime_offset
 #' @export
 methods::setMethod("as.nanotime", methods::signature(from="datetime_offset"),
-    function(from, ..., month = 1, day = 1,
-             hour = 0, minute = 0, second = 0, tz = "GMT") {
+    function(from, ..., year = 1970L, month = 1L, day = 1L,
+             hour = 0L, minute = 0L, second = 0L, nanosecond = 0L, tz = "") {
         x <- from
+        year(x) <- update_missing(year(x), year)
         month(x) <- update_missing(month(x), month)
         day(x) <- update_missing(day(x), day)
         hour(x) <- update_missing(hour(x), hour)
         minute(x) <- update_missing(minute(x), minute)
         second(x) <- update_missing(second(x), second)
-        tz(x) <- ifelse(is.na(tz(x)) & is.na(hour_offset(x)), tz, tz(x))
-        as.nanotime(format(x))
+        nanosecond(x) <- update_missing(nanosecond(x), nanosecond)
+
+        tz(x) <- ifelse(is.na(tz(x)) & is.na(hour_offset(x)), clean_tz(tz), tz(x))
+        as.nanotime(format_ISO8601(x))
 })
 
 update_missing <- function(original, replacement) ifelse(is.na(original), replacement, original)
+update_missing_tz <- function(x, tzone) {
 
-#' @rdname from_datetime_offset
-#' @export
-as.POSIXct.datetime_offset <- function(x, tz = mode_tz(x), ...) {
-    as.POSIXct(as.nanotime(x), tz = tz)
+
 }
 
 #' @rdname from_datetime_offset
 #' @export
-as.POSIXlt.datetime_offset <- function(x, tz = mode_tz(x), ...) {
-    as.POSIXlt(as.nanotime(x), tz = tz)
+as.POSIXct.datetime_offset <- function(x, tz = mode_tz(x), ...,
+                                       year = 1970L, month = 1L, day = 1L,
+                                       hour = 0L, minute = 0L, second = 0L, nanosecond = 0L) {
+    as.POSIXct(as.nanotime(x, ...,
+                           year = year, month = month, day = day,
+                           hour = hour, minute = minute, second = second, nanosecond = nanosecond,
+                           tz = tz),
+               tz = tz)
+}
+
+#' @rdname from_datetime_offset
+#' @export
+as.POSIXlt.datetime_offset <- function(x, tz = mode_tz(x), ...,
+                                       year = 1970L, month = 1L, day = 1L,
+                                       hour = 0L, minute = 0L, second = 0L, nanosecond = 0L) {
+    as.POSIXct(as.nanotime(x, ...,
+                           year = year, month = month, day = day,
+                           hour = hour, minute = minute, second = second, nanosecond = nanosecond,
+                           tz = tz),
+               tz = tz)
 }