+++ tests/testthat/test-from_datetime_offset.r
@@ -1,7 +1,7 @@
 test_that("as.Date()", {
     expect_equal(as.Date(as_datetime_offset("2020")),
                  as.Date("2020-01-01"))
-    expect_equal(as.Date(as_datetime_offset("2020"), 6, 15),
+    expect_equal(as.Date(as_datetime_offset(""), year = 2020, month = 6, day = 15),
                  as.Date("2020-06-15"))
     expect_equal(as.Date(as_datetime_offset("2020-03-23")),
                  as.Date("2020-03-23"))